refactor: use BSBLanData from init
