+++ homeassistant/components/bsblan/diagnostics.py
@@ -7,8 +7,8 @@ from typing import Any
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 
+from . import BSBLanData
 from .const import DOMAIN
-from .models import BSBLanData
 
 
 async def async_get_config_entry_diagnostics(
