+++ homeassistant/components/bsblan/entity.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
+from . import BSBLanData
 from .const import DOMAIN
 from .coordinator import BSBLanUpdateCoordinator
-from .models import BSBLanData
 
 
 class BSBLanEntity(CoordinatorEntity[BSBLanUpdateCoordinator]):
@@ -17,13 +17,8 @@ class BSBLanEntity(CoordinatorEntity[BSBLanUpdateCoordinator]):
 
     def __init__(self, coordinator: BSBLanUpdateCoordinator, data: BSBLanData) -> None:
         """Initialize BSBLan entity."""
-        super().__init__(coordinator)
-        self._data = data
-        host = (
-            coordinator.config_entry.data["host"]
-            if coordinator.config_entry
-            else "unknown"
-        )
+        super().__init__(coordinator, data)
+        host = self.coordinator.config_entry.data["host"]
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, data.device.MAC)},
             name=data.device.name,
@@ -32,8 +27,3 @@ class BSBLanEntity(CoordinatorEntity[BSBLanUpdateCoordinator]):
             sw_version=data.device.version,
             configuration_url=f"http://{host}",
         )
-
-    @property
-    def available(self) -> bool:
-        """Return True if entity is available."""
-        return self.coordinator.last_update_success
